test: migrate Incident[Exporter|Update]IT
 * Migrate tests to MultiDB test, and merge them together.
